# Query: fix streamSeriesSet defer in for loop