--- a/pkg/store/proxy.go
+++ b/pkg/store/proxy.go
@@ -418,28 +418,28 @@ func startStreamSeriesSet(
 				}
 			}
 		}()
-		for {
+		handleRecvResponse := func() (next bool) {
 			frameTimeoutCtx, cancel := frameCtx(s.responseTimeout)
 			defer cancel()
 			var rr *recvResponse
 			select {
 			case <-ctx.Done():
 				s.handleErr(errors.Wrapf(ctx.Err(), "failed to receive any data from %s", s.name), done)
-				return
+				return false
 			case <-frameTimeoutCtx.Done():
 				s.handleErr(errors.Wrapf(frameTimeoutCtx.Err(), "failed to receive any data in %s from %s", s.responseTimeout.String(), s.name), done)
-				return
+				return false
 			case rr = <-rCh:
 			}
 
 			if rr.err == io.EOF {
 				close(done)
-				return
+				return false
 			}
 
 			if rr.err != nil {
 				s.handleErr(errors.Wrapf(rr.err, "receive series from %s", s.name), done)
-				return
+				return false
 			}
 			numResponses++
 			bytesProcessed += rr.r.Size()
@@ -455,8 +455,14 @@ func startStreamSeriesSet(
 				case s.recvCh <- series:
 				case <-ctx.Done():
 					s.handleErr(errors.Wrapf(ctx.Err(), "failed to receive any data from %s", s.name), done)
-					return
+					return false
 				}
+			}
+			return true
+		}
+		for {
+			if !handleRecvResponse() {
+				return
 			}
 		}
 	}()