@hanjm 2021-11-04
> What about the benchmarks? What improvement do they show? 🤔

the `defer` only exec on function return, if we put the `defer cancel` in for loop, then the timer created by context only released on function return, which lead too many timer. so the common best practise is use a func in for loop.
ref: https://golang.org/ref/spec#Defer_statements      https://stackoverflow.com/questions/45617758/defer-in-the-loop-what-will-be-better

ok, let me add a benchmark.🤔