@hanjm 2021-11-04
Hi @GiedriusS , I  make a benchmark. Do i need to add it to `proxy_test.go`? 😄

```bash
// BenchmarkFrameCtxDeferLoop benchmark defer in for loop (#4830).
// go test -bench FrameCtxDefer -run FrameCtxDefer -v
func BenchmarkFrameCtxDeferLoop(b *testing.B) {
	b.ReportAllocs()
	b.RunParallel(func(pb *testing.PB) {
		for pb.Next() {
			for j := 0; j < 1000; j++ {
				_, cancel := frameCtx(time.Second * 5)
				defer cancel()
			}
		}
	})
}

// BenchmarkFrameCtxDeferFunc benchmark defer func in for loop (#4830).
func BenchmarkFrameCtxDeferFunc(b *testing.B) {
	b.ReportAllocs()
	b.RunParallel(func(pb *testing.PB) {
		for pb.Next() {
			for j := 0; j < 1000; j++ {
				func() {
					_, cancel := frameCtx(time.Second * 5)
					defer cancel()
				}()
			}
		}
	})
}
```

```
$ go test -bench FrameCtxDefer -run FrameCtxDefer -v
goos: darwin
goarch: amd64
pkg: github.com/thanos-io/thanos/pkg/store
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkFrameCtxDeferLoop
BenchmarkFrameCtxDeferLoop-12               4972            246576 ns/op          332064 B/op       4998 allocs/op
BenchmarkFrameCtxDeferFunc
BenchmarkFrameCtxDeferFunc-12              12950             87979 ns/op          208000 B/op       4000 allocs/op
PASS
ok      github.com/thanos-io/thanos/pkg/store   4.783s
```