@GiedriusS 2021-11-08
Thank you! BenchmarkProxySeries shows the following improvements:
```
name                                                                                     old time/op    new time/op    delta
ProxySeries/1000000SeriesWith1Samples/4_client_with_1_samples,_250000_series_each-16        1.05s ± 5%     1.00s ± 6%   -4.46%  (p=0.005 n=10+10)
ProxySeries/1000000SeriesWith1Samples/single_client_with_1_samples,_1000000_series-16       1.09s ± 5%     0.94s ± 2%  -14.33%  (p=0.000 n=10+9)
ProxySeries/100000SeriesWith100Samples/4_client_with_25_samples,_25000_series_each-16       106ms ± 3%     100ms ± 7%   -6.18%  (p=0.000 n=10+10)
ProxySeries/100000SeriesWith100Samples/single_client_with_100_samples,_100000_series-16     107ms ± 5%      94ms ± 3%  -12.68%  (p=0.000 n=9+9)
ProxySeries/1SeriesWith10000000Samples/4_client_with_2500000_samples,_1_series_each-16     40.8µs ± 3%    40.4µs ± 4%     ~     (p=0.278 n=10+9)
ProxySeries/1SeriesWith10000000Samples/single_client_with_10000000_samples,_1_series-16    24.4µs ± 5%    23.7µs ± 2%     ~     (p=0.133 n=10+7)

name                                                                                     old alloc/op   new alloc/op   delta
ProxySeries/1000000SeriesWith1Samples/4_client_with_1_samples,_250000_series_each-16        213MB ± 0%     189MB ± 0%  -11.14%  (p=0.000 n=9+9)
ProxySeries/1000000SeriesWith1Samples/single_client_with_1_samples,_1000000_series-16       269MB ± 0%     189MB ± 0%  -29.72%  (p=0.000 n=10+9)
ProxySeries/100000SeriesWith100Samples/4_client_with_25_samples,_25000_series_each-16      19.9MB ± 1%    19.1MB ± 0%   -4.10%  (p=0.000 n=8+10)
ProxySeries/100000SeriesWith100Samples/single_client_with_100_samples,_100000_series-16    22.5MB ± 2%    19.1MB ± 0%  -15.44%  (p=0.000 n=10+10)
ProxySeries/1SeriesWith10000000Samples/4_client_with_2500000_samples,_1_series_each-16     10.8kB ± 0%    10.8kB ± 0%     ~     (all equal)
ProxySeries/1SeriesWith10000000Samples/single_client_with_10000000_samples,_1_series-16    5.02kB ± 0%    5.02kB ± 0%     ~     (all equal)

name                                                                                     old allocs/op  new allocs/op  delta
ProxySeries/1000000SeriesWith1Samples/4_client_with_1_samples,_250000_series_each-16        5.30M ± 0%     5.00M ± 0%   -5.60%  (p=0.000 n=9+9)
ProxySeries/1000000SeriesWith1Samples/single_client_with_1_samples,_1000000_series-16       6.00M ± 0%     5.00M ± 0%  -16.67%  (p=0.000 n=10+9)
ProxySeries/100000SeriesWith100Samples/4_client_with_25_samples,_25000_series_each-16        510k ± 0%      500k ± 0%   -2.00%  (p=0.000 n=8+10)
ProxySeries/100000SeriesWith100Samples/single_client_with_100_samples,_100000_series-16      544k ± 1%      500k ± 0%   -8.00%  (p=0.000 n=10+10)
ProxySeries/1SeriesWith10000000Samples/4_client_with_2500000_samples,_1_series_each-16        177 ± 0%       177 ± 0%     ~     (all equal)
ProxySeries/1SeriesWith10000000Samples/single_client_with_10000000_samples,_1_series-16       106 ± 0%       106 ± 0%     ~     (all equal)
```